Remove boilerplate
Project ID and component ID are accessible via context.

Closes #244.

Fixes #246. There isn't a nicer, more specific way that is still easy to maintain
to link to the currently released version of ODS documentation. On the
on hand that is unfortunate because people need to click on "Jenkins
Shared library", on the other hand it might help that people get
familiar with the documentation which is not bad either.