Remove deprecated dockerImageRepository field
It is not required to work. However, it has a weird side-effect: if the
name of the image stream equals a repository on DockerHub, then
automatic pull-through gets activated. Example: you provision the
be-golang-plain quickstarters with the name "golang". Now the
imagestream will be named "golang", and your DeploymentConfig points to
"golang:latest". OpenShift will pull that image from DockerHub, and
deploy it ... we certainly do not want that, so it is better to remove
this field.

The docs say:

  dockerImageRepository is optional, if specified this stream is backed by a container repository on this server Deprecated: This field is deprecated as of v3.7 and will be removed in a future release. Specify the source for the tags to be imported in each tag via the spec.tags.from reference instead.